repeat code 1
Type: Duplication of code (Grenade and Molotov classes both have an update() method that contains the same piece of code to update the object's movement. This code is responsible for calculating the object's velocity, detecting collisions, applying gravity, etc.)

Possible problems: Duplication of code can cause multiple problems. First, it increases the amount of code, making the code more complex and difficult to maintain and understand. Duplicate code requires changes in multiple places, which can cause consistency issues when modifications are made or bugs are fixed. Furthermore, duplicating code violates the DRY principle (Don't Repeat Yourself, Don't Repeat Yourself), which recommends avoiding duplication of code in a system.
Reason for the problem: In this case, there is a lot of code duplication in the update methods in the Grenade and Molotov classes. It could be that the author of the class copy-pasted that code into both classes, not realizing that a shared method could be used to avoid duplication. It can be caused by poor planning or insufficient attention to code reuse.