Magic Numbers
1. Type: Magic Numbers
2. What does this problem lead to? Using hard-coded values results in less readable code and makes the code difficult to understand and modify. If the same hard-coded value is used in multiple places, modifying these values requires making changes in multiple places in the code, increasing the complexity of maintenance. Additionally, when these values need to be adjusted, a global search and replace for every occurrence of the hardcoded value is required, which is prone to errors.
What is causing this problem? The reason for this problem is that the value is used directly in the code without extracting it as a constant or parameter. The occurrence of hard-coded values may be due to the fact that the configuration and flexibility of values were not considered during the coding process.