Long Function
1. Defect type of "Long Function"
2. What does this problem lead to? Long initialization methods make the code difficult to understand and maintain. When a method contains too much functionality and logic, it increases the difficulty of reading and understanding the code, and also increases the difficulty of modifying and debugging the code. Furthermore, if an error occurs during initialization or a modification is required, it needs to be located and modified in a bulky method, increasing the risk of errors and introducing new bugs.
What is causing this problem? The reason for this problem is to gather too much functionality and logic in one method. It may be that the initial design did not consider the decomposition function, the lack of awareness of code refactoring, or the continuous increase of functions during the development process without timely refactoring, resulting in lengthy initialization methods.